Added team switching functionality

+++ client/src/components/views/HomePlannedTasks.js
@@ -9,13 +9,16 @@ import ReportProblem from "./ReportProblem";
 import EditUncompletedPlan from "./EditUncompletedPlan";
 
 
-function HomePlannedTasks({ refreshHomeTasks, week_start, week_end, prevweek_start, prevweek_end}) {
+function HomePlannedTasks({ refreshHomeTasks, week_start, week_end, prevweek_start, prevweek_end, team_id}) {
   const [progress, setProgress] = useState([]);
   const [students, setStudents] = useState("");
 
   const getProgress = async () => {
     try {
-      const response = await fetch(`http://localhost:5000/1/home/${prevweek_start}/${prevweek_end}/plannedtasks`);
+      if (team_id === "default"){
+        return;
+      }
+      const response = await fetch(`http://localhost:5000/${team_id}/home/${prevweek_start}/${prevweek_end}/plannedtasks`);
       const jsonData = await response.json();
       setProgress(jsonData);
     } catch (err) {
@@ -28,6 +31,7 @@ function HomePlannedTasks({ refreshHomeTasks, week_start, week_end, prevweek_sta
     updateStudents();
     
   }, []);
+  //console.log("progress "+progress);
 
   useEffect(() => {
     updateStudents();
@@ -66,9 +70,12 @@ function HomePlannedTasks({ refreshHomeTasks, week_start, week_end, prevweek_sta
 
   const markAsComplete = async ({plan_id, plan_title, description, student, related_objectives}) =>{
      try {
+      if (team_id === "default"){
+        return;
+      }
        const body = { plan_title, description, student, related_objectives};
        const response = await fetch(
-         `http://localhost:5000/1/progress/${plan_id}/markascomplete/${week_start}/${week_end}`,
+         `http://localhost:5000/${team_id}/progress/${plan_id}/markascomplete/${week_start}/${week_end}`,
          {
            method: "PUT",
            headers: { "Content-Type": "application/json" },
@@ -83,8 +90,11 @@ function HomePlannedTasks({ refreshHomeTasks, week_start, week_end, prevweek_sta
 
    const removePlan = async ({plan_id}) =>{
      try {
+      if (team_id === "default"){
+        return;
+      }
         const response = await fetch(
-            `http://localhost:5000/1/report/plan/${plan_id}`,
+            `http://localhost:5000/{team_id}/report/plan/${plan_id}`,
             {
               method: "DELETE",
               headers: { "Content-Type": "application/json" },
@@ -100,9 +110,12 @@ function HomePlannedTasks({ refreshHomeTasks, week_start, week_end, prevweek_sta
 
    const markUncomplete = async ({plan_id,plan_title}) =>{
      try {
+      if (team_id === "default"){
+        return;
+      }
        const body = { plan_title };
        const response = await fetch(
-         `http://localhost:5000/1/progress/markasincomplete/${plan_id}`,
+         `http://localhost:5000/${team_id}/progress/markasincomplete/${plan_id}`,
          {
            method: "PUT",
            headers: { "Content-Type": "application/json" },
@@ -134,34 +147,34 @@ function HomePlannedTasks({ refreshHomeTasks, week_start, week_end, prevweek_sta
                 {progress.map(({plan_id, plan_title, description, student, related_objectives, marked_complete}, index) => {
                   return (
                     <>
-                    <Draggable key={`compplanid${plan_id}`} draggableId={`compplanid${plan_id}`} index={index}>
+                    <Draggable key={`homeplanid${plan_id}`} draggableId={`homeplanid${plan_id}`} index={index}>
                       {(provided) => (
                         <li ref={provided.innerRef} {...provided.draggableProps} {...provided.dragHandleProps}>
                           <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', width:"100%" }}>
                             <RxDragHandleDots2 className="float-left mr-2"></RxDragHandleDots2>
-                            <EditUncompletedPlan plan_id={plan_id} plan_title={plan_title} plan_description={description} student={student} related_objectives={related_objectives} marked_complete={marked_complete} refreshHomeTasks={refreshHomeTasks}></EditUncompletedPlan>
-                            {marked_complete ? (<></>):(<ReportProblem plan_id={plan_id} week_start={week_start} week_end={week_end}></ReportProblem>) }
+                            <EditUncompletedPlan plan_id={plan_id} plan_title={plan_title} plan_description={description} student={student} related_objectives={related_objectives} marked_complete={marked_complete} refreshHomeTasks={refreshHomeTasks} team_id={team_id}></EditUncompletedPlan>
+                            {marked_complete ? (<></>):(<ReportProblem team_id={team_id} plan_id={plan_id} week_start={week_start} week_end={week_end}></ReportProblem>) }
                             {marked_complete ? (<button className = "btn3 float-right" onClick={() => markUncomplete({plan_id,plan_title})}><AiFillCheckCircle style={{fontSize:'1.25rem'}}></AiFillCheckCircle></button>):(<button className = "btn3 float-right" onClick={() => markAsComplete({plan_id,plan_title, description, student, related_objectives})}><AiOutlineCheckCircle style={{fontSize:'1.25rem'}}></AiOutlineCheckCircle></button>) }
                           <button className = "btn3 float-right" data-toggle="modal" data-target={`#HSDelid${plan_id}`} ><AiOutlineDelete style={{fontSize:'1.25rem'}}></AiOutlineDelete></button>
                           </div>
                         </li>
                       )}
                     </Draggable>
-                    <div class="modal fade" id={`HSDelid${plan_id}`} tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
-                    <div class="modal-dialog modal-dialog-centered" role="document">
-                    <div class="modal-content">
-                        <div class="modal-header">
+                    <div className="modal fade" id={`HSDelid${plan_id}`} tabIndex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
+                    <div className="modal-dialog modal-dialog-centered" role="document">
+                    <div className="modal-content">
+                        <div className="modal-header">
                         <h5 className="modal-title" id="exampleModalLongTitle">Delete Task?</h5>
                         <button type="button" className="close" data-dismiss="modal" aria-label="Close">
                             <span aria-hidden="true">&times;</span>
                         </button>
                         </div>
-                        <div class="modal-body">
+                        <div className="modal-body">
                         Are you sure you want to delete this task?
                         </div>
-                        <div class="modal-footer">
-                        <button type="button" class="btn btn-secondary" data-dismiss="modal">No</button>
-                        <button type="button" class="btn btn-primary" onClick={() => removePlan({plan_id})} data-dismiss="modal">Yes</button>
+                        <div className="modal-footer">
+                        <button type="button" className="btn btn-secondary" data-dismiss="modal">No</button>
+                        <button type="button" className="btn btn-primary" onClick={() => removePlan({plan_id})} data-dismiss="modal">Yes</button>
                         </div>
                     </div>
                     </div>
